modules/mod-011-ml-serving-apis/exercise-03-production-api/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Caching
aiocache==0.12.2

# Rate Limiting
slowapi==0.1.9

# Security
cryptography==41.0.7
itsdangerous==2.1.2

# ML
scikit-learn==1.3.2
numpy==1.26.2

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
